requirements.txt
requests
ipfshttpclient
python-dotenv
ipinfo